requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-cloud-firestore==2.*